chore: update dependencies and enhance development scripts
- Updated package-lock.json and package.json to include new dependencies: deepmerge and inquirer.
- Added VSCode settings for multi-cursor support.
- Enhanced dev.sh script to support 'install' and 'build' commands for Docker.
- Refactored addon handling in addons.ts to streamline action application.
- Introduced applyNginxReplace function for better Nginx configuration management.
- Added new Node 23 addon with Dockerfile and Nginx proxy configuration.
- Updated existing Nginx configurations to use placeholders for dynamic proxy settings.
- Improved error handling in envFile.sh for missing environment files.

+++ bin/dev.sh
@@ -6,6 +6,16 @@ if [ "$1" = 'ssh' ]; then
     exit
 fi;
 
+if [ "$1" = 'install' ]; then
+    docker compose exec -ti builder bash -c "npm install"
+    exit
+fi;
+
+if [ "$1" = 'build' ]; then
+    docker compose exec -ti builder bash -c "npm run build"
+    exit
+fi;
+
 docker compose down
 DOCKER_UID="$(id -u)" DOCKER_GID="$(id -g)" docker compose \
   -f docker-compose.yml -f docker-compose.dev.yml up -d "$@"
